w3c-reports/12092014/efl/runned-tests/vibration/api-is-present-expected.txt

Description

This test checks for the presence of the vibrate() method, taking vendor prefixes into account.

The feature(s) tested here are known to have been made available via vendor prefixes; you can run this test with vendor prefixes enabled.

PASS vibrate() is present on navigator 
